+++ dotnet_project/CfdCsharpProject_Core2/CfdCsharpProject_Core2.csproj
@@ -39,5 +39,6 @@
 
   <ItemGroup>
     <Compile Include="../../src/*.cs" />
+    <Compile Include="../../src/internal/*.cs" />
   </ItemGroup>
 </Project>
