+++ src/CoinUtil.cs
@@ -1,0 +1,6 @@
+using System;
+using System.Runtime.InteropServices;
+
+namespace Cfd
+{
+}
