+++ src/ExtPubkey.cs
@@ -1,0 +1,14 @@
+using System;
+using System.Runtime.InteropServices;
+
+namespace Cfd
+{
+  /// <summary>
+  /// extend key type.
+  /// </summary>
+  public enum CfdExtKeyType
+  {
+    Privkey = 0,  //!< extended privkey
+    Pubkey        //!< extended pubkey
+  };
+}
