+++ src/internal/CTransaction.cs
@@ -1,0 +1,22 @@
+using System;
+using System.Runtime.InteropServices;
+
+namespace Cfd
+{
+  /*
+    internal static class CTransaction
+    {
+      CFDC_API int CfdInitializeMultisigSign(void* handle, void** multisign_handle);
+
+      CFDC_API int CfdAddMultisigSignData(
+          void* handle, void* multisign_handle, const char* signature,
+          const char* related_pubkey);
+
+      CFDC_API int CfdAddMultisigSignDataToDer(
+          void* handle, void* multisign_handle, const char* signature,
+          int sighash_type, bool sighash_anyone_can_pay, const char* related_pubkey);
+
+      CFDC_API int CfdFreeMultisigSignHandle(void* handle, void* multisign_handle);
+    }
+  */
+}
